[actuation] changed to simple actuation model
